+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 7.0.0 (unreleased)
 
 - Add support for PS256, PS384 and PS512
+- Add support for verifying with modulus/exponent components for RSA
 - Change API for both sign/verify to take a `Key` enum rather than bytes
 - Update to 2018 edition